requirements.txt
Flask==2.0.1
flask-cors==3.0.10
sentence-transformers==2.2.0
numpy==1.22.4
pandas==1.3.1
scikit-learn==0.24.2
pymongo==3.11.4
python-dotenv==0.19.0
gunicorn==20.1.0 